refactor: install a single multi-project service; drop port machinery
webmux is one server per machine, so `webmux service install` now creates
one service under a fixed name (`webmux`) instead of one `webmux-<project>`
per project. Add more projects from the dashboard or `webmux project add`;
`webmux project migrate` (next commit) folds legacy per-project units in.

- service.ts: fixed serviceName, no per-service port auto-pick. `--port`
  wins; a bare reinstall reuses the existing unit's port; otherwise 5111.
  Move readPortFromUnit here (still needed for reinstall + `webmux update`).
- Delete install-ports.ts (pickFreePort / discoverTakenPorts /
  readInstalledServicePorts) — the per-service port scan only existed so
  many per-project servers could share nearby ports.
- server.ts: bind the configured port and fail clearly on EADDRINUSE
  (almost always another webmux → the migration cue) instead of walking
  PORT..PORT+100. Drop WEBMUX_PORT_STRICT and the strict/walk split.
- webmux.ts: drop the WEBMUX_PORT_STRICT passthrough + the "falls back to a
  free port" banner.
- service-restart.ts: `webmux update` now also picks up `webmux.service`,
  not just legacy `webmux-*` units.
- Move the readPortFromUnit tests into service.test.ts.

allocateServicePorts (worktree dev-service ports in domain/policies.ts) is
unrelated and untouched.
